fz.enocean_ptm215z, fz.enocean_ptm215ze and fz.enocean_ptm216z
